Module aws_sdk_healthlake::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

Access is denied. Your account is not authorized to perform this operation.

ConflictException

The Data Store is in a transition state and the user requested action can not be performed.

CreateFHIRDatastoreError

DeleteFHIRDatastoreError

DescribeFHIRDatastoreError

DescribeFHIRExportJobError

DescribeFHIRImportJobError

InternalServerException

Unknown error occurs in the service.

ListFHIRDatastoresError

ListFHIRExportJobsError

ListFHIRImportJobsError

ListTagsForResourceError

ResourceNotFoundException

The requested Data Store was not found.

StartFHIRExportJobError

StartFHIRImportJobError

TagResourceError

ThrottlingException

The user has exceeded their maximum number of allowed calls to the given API.

UntagResourceError

ValidationException

The user input parameter was invalid.

Enums

CreateFHIRDatastoreErrorKind

DeleteFHIRDatastoreErrorKind

DescribeFHIRDatastoreErrorKind

DescribeFHIRExportJobErrorKind

DescribeFHIRImportJobErrorKind

ListFHIRDatastoresErrorKind

ListFHIRExportJobsErrorKind

ListFHIRImportJobsErrorKind

ListTagsForResourceErrorKind

StartFHIRExportJobErrorKind

StartFHIRImportJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind