Module aws_sdk_healthlake::model

Modules

datastore_filter

See DatastoreFilter

datastore_properties

See DatastoreProperties

export_job_properties

See ExportJobProperties

import_job_properties

See ImportJobProperties

kms_encryption_config

See KmsEncryptionConfig

preload_data_config

See PreloadDataConfig

s3_configuration

See S3Configuration

sse_configuration

See SseConfiguration

tag

See Tag

Structs

DatastoreFilter

The filters applied to Data Store query.

DatastoreProperties

Displays the properties of the Data Store, including the ID, Arn, name, and the status of the Data Store.

ExportJobProperties

The properties of a FHIR export job, including the ID, ARN, name, and the status of the job.

ImportJobProperties

Displays the properties of the import job, including the ID, Arn, Name, and the status of the Data Store.

KmsEncryptionConfig

PreloadDataConfig

The input properties for the preloaded Data Store. Only data preloaded from Synthea is supported.

S3Configuration

SseConfiguration

Tag

Enums

CmkType

DatastoreStatus

FhirVersion

InputDataConfig

JobStatus

OutputDataConfig

PreloadDataType